refactor: extract util functions to hook utils file
